fix: shrink host.tsx under the file-size cap after the scratch wiring
Extract the optimistic-overlay React binding to
use-optimistic-engine-state.ts and the openTaskWorktree wiring to
open-task-worktree.ts; fold useScratchAdopt inside useScratchShell so the
host consumes one hook. host.tsx 515 → 494 lines.
