+++ packages/kobe/src/tui-react/workspace/open-task-worktree.ts
@@ -1,5 +1,7 @@
 import { existsSync } from "node:fs"
+import { t } from "../../tui/i18n"
 import { detectWorktreeOpener, openWorktree } from "../../tui/lib/worktree-opener"
+import type { Task } from "../../types/task"
 
 type OpenTaskWorktreeDeps = {
   taskPath?: string
@@ -31,4 +33,23 @@ export async function requestTaskWorktreeOpen(id: string, deps: OpenTaskWorktree
   if (!openWorktree(path, opener)) {
     deps.notifyError(deps.openFailedMessage(opener.label))
   }
+}
+
+/** The host's standard wiring for {@link requestTaskWorktreeOpen}: resolve
+ *  the task's recorded path and use the stock toast copy. */
+export function openTaskWorktreeFor(
+  id: string,
+  opts: {
+    tasks: readonly Task[]
+    ensureWorktree: (id: string) => Promise<string>
+    notifyError: (message: string) => void
+  },
+): void {
+  void requestTaskWorktreeOpen(id, {
+    taskPath: opts.tasks.find((task) => task.id === id)?.worktreePath,
+    ensureWorktree: opts.ensureWorktree,
+    notifyError: opts.notifyError,
+    noEditorMessage: t("tasks.toast.noEditor"),
+    openFailedMessage: (label) => t("tasks.toast.openWorktreeFailed", { label }),
+  })
 }
